requirements.txt
# RAG Studio - Python dependencies
# Install with: pip install -r requirements.txt

# Core LangChain
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-huggingface>=0.0.3
langchain-groq>=0.1.6
langchain-google-genai>=1.0.0
langchain-experimental>=0.0.60

# LLM & Embedding providers
groq
google-generativeai>=0.7.0

# Vector store
faiss-cpu>=1.7.4

# Embeddings & Reranking (local models)
sentence-transformers>=2.7.0

# Sparse retrieval
rank-bm25>=0.2.2

# Evaluation frameworks
ragas>=0.1.9
deepeval>=0.21.0
datasets>=2.18.0

# Document loading
pypdf>=4.0.0

# Streamlit
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.26.0
